Add support for keypairs generated by the Rust wallet.
The Helium Rust wallet (`helium-wallet-rs`) duplicates the network
and key type byte (also known as the "tag") when encoding the keypair as a
binary blob. It places the tag byte at the begining of the buffer, where
it is expected by libp2p, but it also adds the byte just before the public
key, where it is not expected by libp2p.

This change accomodates the duplication and allows for keys generated by
the wallet to be loaded by libp2p.